requirements.txt
kafka-python==2.0.2
Faker==20.1.0
pyyaml==6.0.1 